fix: cleaned up code, fixed box shadow on background
