+++ lib/ui/screens/intro/background.dart
@@ -1,0 +1,98 @@
+import 'package:flutter/material.dart';
+
+class IntroBackground extends StatelessWidget {
+  const IntroBackground({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Stack(
+      children: [
+        Container(
+          width: double.infinity,
+          height: double.infinity,
+          decoration: const BoxDecoration(
+            image: DecorationImage(
+                image: AssetImage('assets/introImages/background-pattern.png'),
+                fit: BoxFit.cover),
+          ),
+        ),
+        Container(
+          width: double.infinity,
+          height: double.infinity,
+          color: const Color.fromRGBO(255, 255, 255, 0.9),
+        ),
+        Positioned(
+          bottom: -200,
+          right: -430,
+          child: SizedBox(
+            width: 1700,
+            height: 830,
+            child: CustomPaint(
+              size: Size(200, (200 * 0.5833333333333334).toDouble()),
+              painter: RPSCustomPainter(),
+            ),
+          ),
+        ),
+        Container(
+          alignment: const Alignment(0, 1),
+          child: Column(
+            children: [
+              Padding(
+                padding: const EdgeInsets.fromLTRB(0, 35, 0, 0),
+                child: Container(
+                  width: 50,
+                  height: 50,
+                  decoration: const BoxDecoration(
+                    image: DecorationImage(
+                      image: AssetImage('assets/introImages/barbeat-logo.png'),
+                      fit: BoxFit.cover,
+                    ),
+                  ),
+                ),
+              ),
+              const Text(
+                'BARBEAT',
+                textAlign: TextAlign.center,
+                style: TextStyle(
+                  fontWeight: FontWeight.bold,
+                  fontSize: 64,
+                  color: Colors.black,
+                ),
+              ),
+            ],
+          ),
+        ),
+      ],
+    );
+  }
+}
+
+// white bottom part
+// This was generated by https://shapemaker.web.app/
+class RPSCustomPainter extends CustomPainter {
+  @override
+  void paint(Canvas canvas, Size size) {
+    Paint paint0 = Paint()
+      ..color = Colors.white
+      ..style = PaintingStyle.fill
+      ..strokeWidth = 1;
+
+    Path path0 = Path();
+    path0.moveTo(size.width * 0.5000000, size.height * 0.3571429);
+    path0.quadraticBezierTo(size.width * 0.6397917, size.height * 0.2871429,
+        size.width * 0.7500000, size.height * 0.3614286);
+    path0.lineTo(size.width * 0.7500000, size.height * 0.7857143);
+    path0.lineTo(size.width * 0.5000000, size.height * 0.7842857);
+    path0.quadraticBezierTo(size.width * 0.5000000, size.height * 0.6775000,
+        size.width * 0.5000000, size.height * 0.3571429);
+    path0.close();
+
+    canvas.drawPath(path0, paint0);
+    canvas.drawShadow(path0, Colors.black, 4, false);
+  }
+
+  @override
+  bool shouldRepaint(covariant CustomPainter oldDelegate) {
+    return true;
+  }
+}
